refactor(server): extract shared test utilities for handler and seed tests
Extract createMockViteServer and createMockLogger functions to a shared
test-utils.ts file to eliminate code duplication between handlers.test.ts
and seeds.test.ts.

Changes:
- Create packages/server/src/__tests__/test-utils.ts with shared mocks
- Update handlers.test.ts to import from test-utils.ts
- Update seeds.test.ts to import from test-utils.ts
- Add MockLogger interface export for type safety

Addresses code challenge recommendation for DRY compliance.
